requirements.txt
python-telegram-bot==20.8
requests
redis
cryptography
python-binance
tenacity
python-dotenv
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
APScheduler==3.6.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.2
fakeredis==2.31.1
filelock
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
google-generativeai
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
multidict==6.6.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prometheus_client==0.23.1
psutil
prometheus_flask_exporter==0.23.2
propcache==0.3.2
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.2
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
aiosqlite
pandas
pandas-ta-openbb